added footer web page and unit test cases

+++ Kebab-grill website/src/main/java/com/example/SpringRestaurantswebsite/Repository/CustomerUserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.stereotype.Repository;
 
 
 @Repository
-public interface CustomerUserRepository extends JpaRepository<CustomerUser, Integer> {
+public interface CustomerUserRepository extends JpaRepository<CustomerUser, Long> {
 
      CustomerUser findByEmail(String emailId);
